# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  handler.AssignPermissionsRequest:
    type: object
  handler.AssignUsersRequest:
    type: object
...
  handler.CreateRoleRequest:
    properties:
      description:
        description: 角色描述
        example: 系统管理员，拥有所有权限
        type: string
      display_name:
        description: 显示名称
        example: 管理员
        type: string
      name:
        description: 角色名称
        example: admin
        type: string
    required:
    - display_name
    - name
    type: object
...
  handler.PermissionResponse:
    properties:
      action:
        description: 操作类型
        example: create
        type: string
      created_at:
        description: 创建时间
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        description: 权限描述
        example: 允许创建用户
        type: string
      display_name:
        description: 显示名称
        example: 创建用户
        type: string
      id:
        description: 权限ID
        example: 1
        type: integer
      name:
        description: 权限名称
        example: user:create
        type: string
      resource:
        description: 资源类型
        example: user
        type: string
      status:
        description: 状态：1-启用，0-禁用
        example: 1
        type: integer
      updated_at:
        description: 更新时间
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  handler.RoleResponse:
    properties:
      created_at:
        description: 创建时间
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        description: 角色描述
        example: 系统管理员
        type: string
      display_name:
        description: 显示名称
        example: 管理员
        type: string
      id:
        description: 角色ID
        example: 1
        type: integer
      name:
        description: 角色名称
        example: admin
        type: string
      status:
        description: 状态：1-启用，0-禁用
        example: 1
        type: integer
      updated_at:
        description: 更新时间
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
  handler.UpdateRoleRequest:
    properties:
      description:
        description: 角色描述（可选）
        example: 系统管理员
        type: string
      display_name:
        description: 显示名称（可选）
        example: 管理员
        type: string
      status:
        description: 状态：1-启用，0-禁用（可选）
        example: 1
        type: integer
    type: object
...
  handler.UserResponse:
    properties:
      created_at:
        description: 创建时间
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        description: 用户邮箱
        example: zhangsan@example.com
        type: string
      id:
        description: 用户ID
        example: 1
        type: integer
      name:
        description: 用户姓名
        example: 张三
        type: string
      status:
        description: 用户状态：1-正常，0-禁用
        example: 1
        type: integer
      updated_at:
        description: 更新时间
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  util.Response:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 响应数据（可选）
      error:
        description: 错误详情（可选）
        type: string
      message:
        description: 响应消息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 这是一个 Go 测试项目的 RESTful API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Test API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: 分页获取角色列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.RoleResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
      summary: 获取角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建一个新角色
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/handler.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: 创建角色
      tags:
      - 角色管理
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: 根据角色ID删除角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: 删除角色
      tags:
      - 角色管理
    get:
      consumes:
      - application/json
      description: 根据角色ID获取角色详细信息
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      summary: 获取角色详情
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 根据角色ID更新角色信息（支持部分更新）
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: 更新角色
      tags:
      - 角色管理
  /roles/{id}/permissions:
    delete:
      consumes:
      - application/json
      description: 从角色中移除权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AssignPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: 移除角色的权限
      tags:
      - 角色管理
    get:
      consumes:
      - application/json
      description: 获取角色拥有的所有权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.PermissionResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
      summary: 获取角色的权限列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 为角色分配权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AssignPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: 分配权限给角色
      tags:
      - 角色管理
  /roles/{id}/users:
    delete:
      consumes:
      - application/json
      description: 从角色中移除用户
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AssignUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: 移除角色的用户
      tags:
      - 角色管理
    get:
      consumes:
      - application/json
      description: 获取拥有该角色的所有用户
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.UserResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
      summary: 获取角色的用户列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 为角色分配用户
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AssignUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: 分配用户给角色
      tags:
      - 角色管理
...
schemes:
- http
- https
swagger: "2.0"